requirements.txt
django==3.1.1
djangorestframework==3.11.1
numpy==1.19.1
cmake
dlib
opencv-python
face-recognition
